+++ api/executive/create.php
@@ -1,0 +1,65 @@
+<?php
+// required headers
+header("Access-Control-Allow-Origin: *");
+header("Content-Type: application/json; charset=UTF-8");
+header("Access-Control-Allow-Methods: POST");
+header("Access-Control-Max-Age: 3600");
+header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With");
+  
+// get database connection
+include_once '../config/database.php';
+  
+// instantiate product object
+include_once '../objects/executives.php';
+  
+$database = new Database();
+$db = $database->getConnection();
+  
+$executive = new Executive($db);
+  
+// get posted data
+$data = json_decode(file_get_contents("php://input"));
+  
+// make sure data is not empty
+if(
+    !empty($data->EUCID) &&
+    !empty($data->Date_elected) &&
+    !empty($data->Dname) 
+){
+  
+    // set product property values
+    $executive->EUCID = $data->EUCID;
+    $executive->Date_elected = $data->Date_elected;
+    $executive->Dname = $data->Dname;
+  
+    // create the product
+    if($executive->create()){
+  
+        // set response code - 201 created
+        http_response_code(201);
+  
+        // tell the user
+        echo json_encode(array("message" => "Executive was created."));
+    }
+  
+    // if unable to create the product, tell the user
+    else{
+  
+        // set response code - 503 service unavailable
+        http_response_code(503);
+  
+        // tell the user
+        echo json_encode(array("message" => "Unable to create executive. Check that member exists."));
+    }
+}
+  
+// tell the user data is incomplete
+else{
+  
+    // set response code - 400 bad request
+    http_response_code(400);
+  
+    // tell the user
+    echo json_encode(array("message" => "Unable to create executive. Data is incomplete."));
+}
+?>